Move ext function utilities into another file
