+++ tavern/_plugins/rest/request.py
@@ -14,9 +14,9 @@ from requests.cookies import cookiejar_from_dict
 from requests.utils import dict_from_cookiejar
 
 from tavern.request.base import BaseRequest
-from tavern.schemas.extensions import get_wrapped_create_function
 from tavern.util import exceptions
 from tavern.util.dict_util import check_expected_keys, deep_dict_merge, format_keys
+from tavern.util.extfunctions import get_wrapped_create_function
 
 logger = logging.getLogger(__name__)
