+++ tavern/schemas/extensions.py
@@ -1,5 +1,3 @@
-import functools
-import importlib
 import logging
 import os
 import re
@@ -8,18 +6,9 @@ from pykwalify.types import is_bool, is_float, is_int
 
 from tavern.util import exceptions
 from tavern.util.exceptions import BadSchemaError
+from tavern.util.extfunctions import import_ext_function
 from tavern.util.loader import ApproxScalar, BoolToken, FloatToken, IntToken
 from tavern.util.strict_util import StrictLevel
-
-
-def _getlogger():
-    """Get logger for this module
-
-    Have to do it like this because the way that pykwalify load extension
-    modules means that getting the logger the normal way just result sin it
-    trying to get the root logger which won't log correctly
-    """
-    return logging.getLogger("tavern.schemas.extensions")
 
 
 # To extend pykwalify's type validation, extend its internal functions
@@ -54,94 +43,6 @@ def validator_like(validate, description):
 int_variable = validator_like(is_int_like, "int-like")
 float_variable = validator_like(is_float_like, "float-like")
 bool_variable = validator_like(is_bool_like, "bool-like")
-
-
-def import_ext_function(entrypoint):
-    """Given a function name in the form of a setuptools entry point, try to
-    dynamically load and return it
-
-    Args:
-        entrypoint (str): setuptools-style entrypoint in the form
-            module.submodule:function
-
-    Returns:
-        function: function loaded from entrypoint
-
-    Raises:
-        InvalidExtFunctionError: If the module or function did not exist
-    """
-    logger = _getlogger()
-
-    try:
-        module, funcname = entrypoint.split(":")
-    except ValueError as e:
-        msg = "Expected entrypoint in the form module.submodule:function"
-        logger.exception(msg)
-        raise exceptions.InvalidExtFunctionError(msg) from e
-
-    try:
-        module = importlib.import_module(module)
-    except ImportError as e:
-        msg = "Error importing module {}".format(module)
-        logger.exception(msg)
-        raise exceptions.InvalidExtFunctionError(msg) from e
-
-    try:
-        function = getattr(module, funcname)
-    except AttributeError as e:
-        msg = "No function named {} in {}".format(funcname, module)
-        logger.exception(msg)
-        raise exceptions.InvalidExtFunctionError(msg) from e
-
-    return function
-
-
-def get_wrapped_response_function(ext):
-    """Wraps a ext function with arguments given in the test file
-
-    This is similar to functools.wrap, but this makes sure that 'response' is
-    always the first argument passed to the function
-
-    Args:
-        ext (dict): $ext function dict with function, extra_args, and
-            extra_kwargs to pass
-
-    Returns:
-        function: Wrapped function
-    """
-    args = ext.get("extra_args") or ()
-    kwargs = ext.get("extra_kwargs") or {}
-    try:
-        func = import_ext_function(ext["function"])
-    except KeyError:
-        raise exceptions.BadSchemaError(
-            "No function specified in external function block"
-        )
-
-    @functools.wraps(func)
-    def inner(response):
-        return func(response, *args, **kwargs)
-
-    inner.func = func
-
-    return inner
-
-
-def get_wrapped_create_function(ext):
-    """Same as above, but don't require a response"""
-    args = ext.get("extra_args") or ()
-    kwargs = ext.get("extra_kwargs") or {}
-    func = import_ext_function(ext["function"])
-
-    @functools.wraps(func)
-    def inner():
-        result = func(*args, **kwargs)
-        _getlogger().info("Result of calling '%s': '%s'", func, result)
-        return result
-
-    inner.func = func
-
-    return inner
 
 
 def _validate_one_extension(input_value):
@@ -406,7 +307,7 @@ def validate_timeout_tuple_or_float(value, rule_obj, path):
     err_msg = "'timeout' must be either a float/int or a 2-tuple of floats/ints - got '{}' (type {})".format(
         value, type(value)
     )
-    logger = _getlogger()
+    logger = logging.getLogger("tavern.schemas.extensions")
 
     def check_is_timeout_val(v):
         if v is True or v is False or not (is_float_like(v) or is_int_like(v)):
